+++ Assets/Scripts/Application Layer/TownSystem/Interface/IShopNPC.cs
@@ -4,6 +4,6 @@ using UnityEngine;
 public interface IShopNPC
 {
     public Transform npcTransform { get; }
-    public long currentMoney { get; }
+    public int currentMoney { get; }
     public event Action ShopMoneyChangedEvent;
 }
