+++ Assets/Scripts/Application Layer/UnitSystem/Interface/IMoneyData.cs
@@ -2,9 +2,9 @@ using UnityEngine;
 
 public interface IMoneyData
 {
-    long money { get; }
-    long carrot { get; }
-    long sunEssence { get; }
-    long moonEssence { get; }
-    long lightningEssence { get; }
+    int money { get; }
+    int carrot { get; }
+    int sunEssence { get; }
+    int moonEssence { get; }
+    int lightningEssence { get; }
 }
